refactor(collections): split table view into modules and refresh row design
The 1375-line monolithic collection_table_view.dart is broken into a
focused module under collection_table/ — one file per role (the view,
the header, the row, the column enum, four cell types). The visual
chrome is dropped: each row is a faint rounded card that floats on the
page, the header sits above as a plain label strip, and the whole table
scrolls together with the collection hero instead of having its own
internal scroll.

Columns are reordered: tag moved to the trailing slot, platform follows
the name. Name (flex 5) and tag (flex 2) are the only stretchy columns;
the rest are fixed width with centred content. Rating renders an em-dash
when unset. Minimum table width before horizontal scrolling kicks in
rose from 600 to 820.

All Cyrillic dartdocs in the touched files (CollectionItem, CachedImage,
CollectionItemsView) were translated to English per the project comment
policy.
